+++ assemblyadapter-recycler/src/main/java/com/github/panpf/assemblyadapter/recycler/AssemblyRecyclerListAdapter.kt
@@ -123,11 +123,6 @@ open class AssemblyRecyclerListAdapter<DATA>
         require(itemFactoryList.isNotEmpty()) { "itemFactoryList Can not be empty" }
     }
 
-    override fun getItemId(position: Int): Long {
-        // todo 不重写 getItemId
-        return position.toLong()
-    }
-
     override fun getItemViewType(position: Int): Int {
         val data = getItem(position) ?: Placeholder
         return itemFactoryStorage.getItemTypeByData(