+++ assemblyadapter-recycler/src/main/java/com/github/panpf/assemblyadapter/recycler/AssemblySingleDataRecyclerListAdapter.kt
@@ -137,11 +137,6 @@ open class AssemblySingleDataRecyclerListAdapter<DATA : Any> :
         super.submitList(list, commitCallback)
     }
 
-    override fun getItemId(position: Int): Long {
-        // todo 不重写 getItemId
-        return position.toLong()
-    }
-
     override fun getItemViewType(position: Int): Int {
         val count = itemCount
         if (position < 0 || position >= count) {