+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * change: The expandable related classes in the list module are moved to the expandable directory
 * change: When the position parameter of AssemblySingleData\Adapter related methods exceeds the range of 0 to count, an IndexOutOfBoundsException will be thrown.
 * new: Assembly\*ExpandableListAdapter added getItemFactoryByChildPosition method
+* change: Assembly\*RecyclerAdapter no longer rewrite the getItemId method
 
 
 # v4.0.0-beta01